feat(sdk): phase 1 bridges — move PolicyAwareToolExecutor to bamboo-tools, add infer_provider/resolve_model
S1.A: move PolicyAwareToolExecutor (+9 tests) from bamboo-server to
bamboo-tools/src/policy_aware.rs; export from lib; leave a re-export shim at
bamboo-server/src/tools/policy_aware.rs.

S1.B: add pure infer_provider() and resolve_model() helpers in
bamboo-engine/src/model_config_helper.rs honoring precedence
model_ref > tier > fallback chain; keep resolve_subagent_model* untouched.
Add tests S-T1.1 (infer_provider mapping) and S-T1.2 (resolve_model precedence).
